Rename package "nodes" to "node"
