+++ src/main/java/guru/franz/mc/bcl/config/node/MaxChunksAmountNode.java
@@ -1,4 +1,4 @@
-package guru.franz.mc.bcl.config.nodes;
+package guru.franz.mc.bcl.config.node;
 
 import ninja.leaping.configurate.objectmapping.serialize.ConfigSerializable;
